+++ src/main/resources/org/jenkinsci/plugins/github_branch_source/AbstractGitHubSCMSource/config-detail.jelly
