+++ src/main/resources/org/jenkinsci/plugins/github_branch_source/Endpoint/config.jelly
