+++ src/main/resources/org/jenkinsci/plugins/github_branch_source/GitHubSCMNavigator/newInstanceDetail.properties
